댓글 수정, 삭제 개발

+++ src/main/java/practice/springmvc/domain/board/comment/CommentService.java
@@ -6,6 +6,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import practice.springmvc.domain.board.Board;
 import practice.springmvc.domain.board.repository.jpa.SpringDataJpaBoardRepository;
+import practice.springmvc.dto.request.CommentRequest;
 import practice.springmvc.utils.SHA256Util;
 
 import java.util.List;
@@ -25,18 +26,26 @@ public class CommentService {
         return commentRepository.save(comment);
     }
 
-    public void update(Long commentId, Comment updateParam) {
-        Comment comment = commentRepository.findById(commentId).orElseThrow();
+    public Comment update(Comment comment, CommentRequest update) {
         Board board = boardRepository.findById(comment.getBoard().getId()).orElseThrow();
-        comment.setNickname(updateParam.getNickname());
-        comment.setPassword(updateParam.getPassword());
-        comment.setIp(updateParam.getIp());
-        comment.setContent(updateParam.getContent());
+        comment.setNickname(update.getNickname());
+        comment.setPassword(update.getPassword());
+        comment.setContent(update.getContent());
         comment.setBoard(board);
+
+        return comment;
     }
 
-    public void delete(Long commentId) {
-        commentRepository.deleteById(commentId);
+    public void delete(Long commentId, String password) {
+        String encryptPwd = encryptSHA256(password);
+        Comment comment = commentRepository.findByIdAndPassword(commentId, encryptPwd);
+
+        if (comment != null) {
+            commentRepository.deleteById(commentId);
+        } else {
+            log.error("deleteId Error {}", comment);
+            throw new RuntimeException("deleteId Error id 삭제 메서드 확인이 필요합니다. info : " + comment);
+        }
     }
 
     public List<Comment> getAllCommentsByBoard(Long boardId) {
